refactor(timetree): extract inference, optimization, and utils to top-level timetree/ domain module
- Move inference/ (backward_pass, forward_pass, branch_length_likelihood, runner) with tests and fixtures
- Move optimization/ (polytomy, relaxed_clock, clock_filter, reroot) with tests
- Move utils.rs (initialize_node_divergences, create_poisson_branch_distributions, extract_node_times)
- Update all imports from crate::commands::timetree:: to crate::timetree::
- commands/timetree/ retains args, run, output, convergence, initialization, refinement (CLI-coupled)
