+++ packages/treetime/src/commands/timetree/initialization.rs
@@ -2,7 +2,6 @@ use crate::alphabet::alphabet::Alphabet;
 use crate::clock::date_constraints::load_date_constraints;
 use crate::commands::shared::args::BranchLengthMode;
 use crate::commands::timetree::args::TreetimeTimetreeArgs;
-use crate::commands::timetree::utils::initialize_node_divergences;
 use crate::gtr::get_gtr::{GtrModelName, get_gtr_by_name, log_gtr, write_gtr_json};
 use crate::make_error;
 use crate::make_report;
@@ -15,6 +14,7 @@ use crate::representation::payload::timetree::EdgeTimetree;
 use crate::representation::payload::timetree::NodeTimetree;
 use crate::seq::alignment::get_common_length;
 use crate::seq::gap_fill::apply_gap_fill;
+use crate::timetree::utils::initialize_node_divergences;
 use eyre::{Report, WrapErr};
 use log::info;
 use parking_lot::RwLock;
