+++ packages/treetime/src/commands/timetree/run.rs
@@ -10,17 +10,13 @@ use crate::coalescent::skyline::{SkylineParams, optimize_skyline};
 use crate::commands::shared::args::BranchLengthMode;
 use crate::commands::timetree::args::{TimeMarginalMode, TreetimeTimetreeArgs};
 use crate::commands::timetree::convergence::metrics::{IterationContext, TimetreeOptimizer};
-use crate::commands::timetree::inference::runner::run_timetree;
 use crate::commands::timetree::initialization::{InputData, initialize_partitions, load_input_data};
-use crate::commands::timetree::optimization::clock_filter::{apply_outlier_bad_branches, report_bad_branches};
-use crate::commands::timetree::optimization::reroot::reroot_tree;
 use crate::commands::timetree::output::auspice::write_auspice_json;
 use crate::commands::timetree::output::confidence::{
   NodeConfidenceInterval, compute_rate_susceptibility, determine_rate_std, extract_confidence_intervals,
   write_confidence_intervals,
 };
 use crate::commands::timetree::refinement::run_refinement_iteration;
-use crate::commands::timetree::utils::initialize_clock_totals_from_time_distributions;
 use crate::optimize::args::BranchOptMethod;
 use crate::optimize::iteration::{apply_damping, save_branch_lengths};
 use crate::optimize::optimize_unified::{run_optimize_mixed, run_optimize_mixed_inner};
@@ -30,6 +26,10 @@ use crate::representation::payload::ancestral::annotate_branch_mutations;
 use crate::representation::payload::timetree::EdgeTimetree;
 use crate::representation::payload::timetree::NodeTimetree;
 use crate::seq::alignment::get_common_length;
+use crate::timetree::inference::runner::run_timetree;
+use crate::timetree::optimization::clock_filter::{apply_outlier_bad_branches, report_bad_branches};
+use crate::timetree::optimization::reroot::reroot_tree;
+use crate::timetree::utils::initialize_clock_totals_from_time_distributions;
 use crate::{make_error, make_report};
 use eyre::{Report, WrapErr};
 use log::{debug, info, warn};
