+++ packages/treetime/src/timetree/inference/__tests__/test_backward_pass.rs
@@ -1,9 +1,9 @@
 #[cfg(test)]
 mod tests {
-  use crate::commands::timetree::inference::backward_pass::propagate_distributions_backward;
   use crate::pretty_assert_ulps_eq;
   use crate::representation::payload::timetree::{EdgeTimetree, NodeTimetree};
   use crate::test_utils::find_node_key_by_name;
+  use crate::timetree::inference::backward_pass::propagate_distributions_backward;
   use eyre::Report;
   use std::sync::Arc;
   use treetime_distribution::Distribution;
