+++ packages/treetime/src/timetree/inference/__tests__/test_branch_length_likelihood.rs
@@ -1,7 +1,7 @@
 #[cfg(test)]
 mod tests {
-  use crate::commands::timetree::inference::branch_length_likelihood::compute_branch_length_distribution;
   use crate::representation::partition::optimization_contribution::OptimizationContribution;
+  use crate::timetree::inference::branch_length_likelihood::compute_branch_length_distribution;
   use approx::assert_abs_diff_eq;
   use eyre::Report;
   use rstest::rstest;
