+++ packages/treetime/src/timetree/inference/__tests__/test_gm_runner/test_gm_runner_marginal_dense.rs
@@ -7,15 +7,15 @@ mod tests {
   use crate::clock::clock_regression::{ClockParams, estimate_clock_model_with_reroot};
   use crate::clock::date_constraints::load_date_constraints;
   use crate::clock::find_best_root::params::BranchPointOptimizationParams;
-  use crate::commands::timetree::inference::runner::run_timetree;
-  use crate::commands::timetree::utils::{
-    extract_node_times, initialize_clock_totals_from_time_distributions, initialize_node_divergences,
-  };
   use crate::gtr::get_gtr::{JC69Params, jc69};
   use crate::representation::partition::marginal_dense::PartitionMarginalDense;
   use crate::representation::partition::timetree::GraphTimetree;
   use crate::representation::partition::traits::PartitionTimetreeAll;
   use crate::representation::payload::timetree::{EdgeTimetree, NodeTimetree};
+  use crate::timetree::inference::runner::run_timetree;
+  use crate::timetree::utils::{
+    extract_node_times, initialize_clock_totals_from_time_distributions, initialize_node_divergences,
+  };
   use eyre::Report;
   use maplit::btreemap;
   use parking_lot::RwLock;
